requirements.txt
# Core Application Dependencies
pandas==2.2.2
openpyxl==3.1.2
Pillow==10.4.0
paramiko==3.4.0
pyserial==3.5
ping3==4.0.4

# Supporting Libraries
numpy==1.26.3
et-xmlfile==1.1.0
lxml==6.0.4
XlsxWriter==3.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.4
six==1.16.0
cryptography==41.0.7
PyNaCl==1.5.0
bcrypt==4.1.2
pycparser==2.21
cffi==2.0.0

# GUI Support
psutil==7.2.2
pywin32==311
wexpect==4.0.0
pexpect==4.9.0
keyring==24.1.1

# Terminal/Documentation Support
docx==0.2.4

# Deployment & Building
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
altgraph==0.17.4
macholib==1.16.3
pefile==2023.2.7
setuptools==69.0.3
packaging==23.2

# Testing & Development
pytest==9.0.3
colorama==0.4.6
iniconfig==2.0.0
pluggy==1.6.0
Pygments==2.20.0
